relative path formatting
added file "find_parent" as first import. This will find the position in the directory tree and allow for relative imports. Must be imported as first module in order to rebase the paths.
Imports in MAFormatter show the different uses of relative imports that will allow for use when a file from a parent is wanted the import is *from* folder *import* file.
When a definition is wanted use *from* folder.file *import* definition

To make code more readable you can also use the *as* key:
*from* folder.file *import* super_long_definintion *as* my_func
You can then use the key "my_func" for the rest of the file.